requirements.txt
annotated-types==0.7.0
anyio==4.9.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
colorlog==6.9.0
cssselect2==0.8.0
distro==1.9.0
fonttools==4.57.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
logging==0.4.9.6
Markdown==3.8
openai==1.70.0
pillow==11.1.0
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
pydyf==0.11.0
pyphen==0.17.2
python-dotenv==1.1.0
python-telegram-bot==22.0
sniffio==1.3.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
weasyprint==65.0
webencodings==0.5.1
zopfli==0.2.3.post1
